+++ AxonsERP.Entities/Models/TaxRateControl.cs
@@ -5,12 +5,12 @@ namespace AxonsERP.Entities.Models
     {
         public string? taxCode { get; set; }
         public DateTime effectiveDate { get; set; }
-        public int rate { get; set; }
+        public float rate { get; set; }
         public string? owner { get; set; }
         public DateTime createDate { get; set; }
         public DateTime lastUpdateDate { get; set; }
         public string? function { get; set; }
-        public int rateOriginal { get; set; }
+        public float rateOriginal { get; set; }
         public string? desc1 { get; set; }
         public string? desc2 { get; set; }
     }
