+++ AxonsERP.Entities/RequestFeatures/TaxRateControlForCreate.cs
@@ -9,8 +9,8 @@ namespace AxonsERP.Entities.RequestFeatures
         [Required]
         public DateTime effectiveDate { get; set; }
 
-        public int rate { get; set; }
+        public float rate { get; set; }
         public string? owner { get; set; }
-        public int rateOriginal { get; set; }
+        public float rateOriginal { get; set; }
     }
 }
